Update cocoplot.py
np.float fix

+++ Python/cocopy/cocoplot.py
@@ -132,8 +132,9 @@ def filter(wavelengths, filtername, rgb_pos ='default', plot=False):
 
         #mean, sigma
         for i in range(3):
-            std = np.float(rgb_pos[i][1])
-            mn  = np.float(rgb_pos[i][0])
+            std = 
+            (rgb_pos[i][1])
+            mn  = float(rgb_pos[i][0])
             c = 1./(std * np.sqrt(2*np.pi))
 
             num = np.arange(len(filter[:,0]))
